@@ -12,10 +12,11 @@
     "ionic:serve": "ionic-app-scripts serve"
   },
   "dependencies": {
+    "@angular/cli": "^8.3.29",
     "@angular/common": "6.1.10",
     "@angular/compiler": "6.1.10",
     "@angular/compiler-cli": "6.1.10",
-    "@angular/core": "^6.1.10",
+    "@angular/core": "^8.2.14",
     "@angular/forms": "6.1.10",
     "@angular/http": "6.1.10",
     "@angular/platform-browser": "6.1.10",
@@ -23,6 +24,8 @@
     "@ionic-native/core": "4.4.0",
     "@ionic-native/splash-screen": "^4.20.0",
     "@ionic-native/status-bar": "4.4.0",
+    "@ionic/angular": "^6.2.6",
+    "@ionic/angular-toolkit": "^7.0.0",
     "@ionic/pro": "1.0.16",
     "@ionic/storage": "2.1.3",
     "@types/jquery": "^3.2.18",
@@ -34,6 +37,7 @@
     "ionicons": "3.0.0",
     "jquery": "^3.2.1",
     "rxjs": "^6.6.7",
+    "rxjs-tslint": "^0.1.8",
     "sw-toolbox": "3.6.0",
     "sweetalert2": "^7.6.2",
     "zone.js": "0.8.29"
@@ -41,7 +45,7 @@
   "devDependencies": {
     "@ionic/app-scripts": "3.1.5",
     "cordova-plugin-splashscreen": "^4.0.3",
-    "typescript": "2.9.2"
+    "typescript": "^3.4.5"
   },
   "description": "An Ionic project",
   "cordova": {
