@@ -1,37 +1,26 @@
-import { BrowserModule } from '@angular/platform-browser';
-import { ErrorHandler, NgModule } from '@angular/core';
-import { IonicApp, IonicErrorHandler, IonicModule } from 'ionic-angular';
-import { SplashScreen } from '@ionic-native/splash-screen';
-import { StatusBar } from '@ionic-native/status-bar';
-import * as $ from 'jquery';
-import { MyApp } from './app.component';
-import { HomePage } from '../pages/home/home';
-import { AuthProvider } from '../providers/auth/auth';
-import { PruebaProvider } from '../providers/prueba/prueba';
-import { HttpModule } from '@angular/http';
-import { HttpClientModule } from '@angular/common/http'
+import { BrowserModule } from "@angular/platform-browser";
+import { ErrorHandler, NgModule } from "@angular/core";
+import { IonicApp, IonicErrorHandler, IonicModule } from "ionic-angular";
+import { SplashScreen } from "@ionic-native/splash-screen";
+import { StatusBar } from "@ionic-native/status-bar";
+import * as $ from "jquery";
+import { MyApp } from "./app.component";
+import { HomePage } from "../pages/home/home";
+import { AuthProvider } from "../providers/auth/auth";
+import { PruebaProvider } from "../providers/prueba/prueba";
+import { HttpClientModule } from "@angular/common/http";
 ///
 @NgModule({
-  declarations: [
-    MyApp,
-    HomePage
-  ],
-  imports: [
-    BrowserModule,
-    IonicModule.forRoot(MyApp),
-     HttpClientModule
-  ],
+  declarations: [MyApp, HomePage],
+  imports: [BrowserModule, IonicModule.forRoot(MyApp), HttpClientModule],
   bootstrap: [IonicApp],
-  entryComponents: [
-    MyApp,
-    HomePage
-  ],
+  entryComponents: [MyApp, HomePage],
   providers: [
     StatusBar,
     SplashScreen,
-    {provide: ErrorHandler, useClass: IonicErrorHandler},
+    { provide: ErrorHandler, useClass: IonicErrorHandler },
     AuthProvider,
-    PruebaProvider
-  ]
+    PruebaProvider,
+  ],
 })
 export class AppModule {}
